Move reflected header exclusion base test to header injection file
