+++ tests/appsec/iast/utils.py
@@ -394,43 +394,3 @@ class BaseTestCookieNameFilter:
 
         meta_req3 = _get_span_meta(self.req3)
         assert "_dd.iast.json" not in meta_req3
-
-
-class BaseTestHeaderInjectionReflectedExclusion:
-    origin_header: None
-    reflected_header: None
-    headers: None
-
-    exclusion_request: None
-    no_exclusion_request: None
-
-    def setup_no_exclusion(self):
-        assert self.origin_header is not None, f"Please set {self}.origin_header"
-        assert isinstance(self.origin_header, str), f"Please set {self}.origin_header"
-        assert self.reflected_header is not None, f"Please set {self}.reflected_header"
-        assert isinstance(self.reflected_header, str), f"Please set {self}.reflected_header"
-
-        self.no_exclusion_request = weblog.get(
-            path="/iast/header_injection/reflected/no-exclusion",
-            params={"origin": self.origin_header, "reflected": self.reflected_header},
-        )
-
-    def test_no_exclusion(self):
-        assert_iast_vulnerability(
-            request=self.no_exclusion_request, vulnerability_count=1, vulnerability_type="HEADER_INJECTION",
-        )
-
-    def setup_exclusion(self):
-        assert self.origin_header is not None, f"Please set {self}.origin_header"
-        assert isinstance(self.origin_header, str), f"Please set {self}.origin_header"
-        assert self.reflected_header is not None, f"Please set {self}.reflected_header"
-        assert isinstance(self.reflected_header, str), f"Please set {self}.reflected_header"
-
-        self.exclusion_request = weblog.get(
-            path="/iast/header_injection/reflected/exclusion",
-            params={"origin": self.origin_header, "reflected": self.reflected_header},
-            headers=self.headers,
-        )
-
-    def test_exclusion(self):
-        BaseSinkTest.assert_no_iast_event(self.exclusion_request)
